+++ bin/run-selenium-desktop.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-java -Dwebdriver.chrome.driver="./chromedriver" -jar ./selenium-server-standalone.jar
+java -Dwebdriver.chrome.driver="/root/instagram-bot.js/bin/chromedriver" -jar /root/instagram-bot.js/bin/selenium-server-standalone.jar